Some refactors; console log text updates
