+++ service/src/utils/create-node.ts
@@ -1,7 +1,7 @@
 import { createLibp2p, Libp2p, Libp2pOptions } from "libp2p";
 import { archLogger } from "./chalk-theme";
+import { setupNodeEventListeners } from "./node-event-listeners";
 
-const idTruncateLimit = 5;
 
 /**
  *
@@ -14,45 +14,15 @@ export async function createNode(
   configOptions: Libp2pOptions,
   connectCallback: () => void,
 ): Promise<Libp2p> {
-  const node = await createLibp2p(configOptions)
+  const node = await createLibp2p(configOptions);
   setupNodeEventListeners(node, name, connectCallback);
 
   const peerId = node.peerId.toString();
-  archLogger.info(`${name} starting with id: ${peerId.slice(peerId.length - idTruncateLimit)}`)
+  archLogger.info(`${name} starting...`);
 
   await node.start();
+  archLogger.notice(`Node started with id: ${peerId.slice(peerId.length - 5)}`);
+
   return node;
 }
 
-const discoverPeers: string[] = [];
-
-function setupNodeEventListeners(
-  node: Libp2p,
-  name: string,
-  connectCallback: () => void,
-) {
-  node.addEventListener('peer:discovery', (evt) => {
-    const peerId = evt.detail.id.toString();
-
-    if (discoverPeers.find((p) => p === peerId) === undefined) {
-      discoverPeers.push(peerId);
-      archLogger.info(`${name}: discovered ${peerId.slice(peerId.length - idTruncateLimit)}`)
-    }
-  })
-
-  node.connectionManager.addEventListener('peer:connect', async (evt) => {
-    const peerId = evt.detail.remotePeer.toString()
-    archLogger.info(`${name}: Connection established to ${peerId.slice(peerId.length - idTruncateLimit)}`);
-
-    connectCallback();
-  })
-
-  node.connectionManager.addEventListener('peer:disconnect', (evt) => {
-    const peerId = evt.detail.remotePeer.toString()
-    archLogger.info(`${name}: Connection dropped from ${peerId.slice(peerId.length - idTruncateLimit)}`)
-  })
-
-  node.pubsub.addEventListener("message", (evt) => {
-    archLogger.info(`${this.name} received a msg: ${new TextDecoder().decode(evt.detail.data)}`)
-  })
-}
